extract build-relevant function to build.py; add build command
