+++ .gitignore
@@ -4,5 +4,5 @@ sqlancer-logs/
 __pycache__/
 **/__pycache__/
 
-
+Dockerfile
 .env
