+++ test.py
@@ -6,31 +6,9 @@ import time
 import shlex
 from datetime import datetime
 
-def load_json(path):
-    with open(path) as f:
-        return json.load(f)
-
 def run_command(cmd, **kwargs):
     print("[CMD]", " ".join(cmd))
     subprocess.run(cmd, check=True, **kwargs)
-
-def ensure_sqlancer_image(force_rebuild=False):
-    if force_rebuild:
-        print("[INFO] Rebuilding SQLancer image due to --cache not specified...")
-        run_command(["docker", "build", "--no-cache", "-t", "sqlancer:latest", "./sqlancer"])
-        return
-
-    result = subprocess.run(
-        ["docker", "images", "--format", "{{.Repository}}:{{.Tag}}"],
-        capture_output=True, text=True
-    )
-    images = result.stdout.strip().splitlines()
-    if "sqlancer:latest" in images:
-        print("[INFO] SQLancer image already exists: sqlancer:latest")
-    else:
-        print("[INFO] SQLancer image not found. Building from ./sqlancer...")
-        run_command(["docker", "build", "-t", "sqlancer:latest", "./sqlancer"])
-
 
 def container_exists(name):
     result = subprocess.run(
@@ -39,12 +17,7 @@ def container_exists(name):
     )
     return name in result.stdout.strip().splitlines()
 
-def start_db_container(dbms, config_path):
-    if not os.path.exists(config_path):
-        print(f"[ERROR] config.json not found: {config_path}")
-        sys.exit(1)
-
-    cfg = load_json(config_path)
+def start_db_container(dbms, cfg):
     image = cfg.get("image")
     container_name = cfg.get("container_name", f"{dbms}-sqlancer")
     env_dict = cfg.get("env", {})
@@ -113,35 +86,13 @@ def start_sqlancer_container(dbms, host_container_name, username, password, orac
         "sqlancer:latest"
     ])
 
-def build_environment(dbms, config_path, use_cache, custom=False, dockerfile_path=""):
-    ensure_network_exists()
-    ensure_sqlancer_image(force_rebuild=not use_cache)
-    cfg = load_json(config_path)
 
-    image = cfg["image"]
-    container_name = cfg["container_name"]
+def test_single(dbms, cfg, use_cache=False):
 
-    if not use_cache and not custom:
-        print(f"[INFO] Pulling image {image}")
-        run_command(["docker", "pull", image])
-
-    if custom:
-        build_cmd = ["docker", "build", "-t", cfg["image"], os.path.dirname(dockerfile_path)]
-        if not use_cache:
-            build_cmd.insert(2, "--no-cache")
-        run_command(build_cmd)
-
-    start_db_container(dbms, config_path)
-
-    return cfg, container_name
-
-
-def test_single(dbms, config_path, use_cache=False):
-    cfg, container_name = build_environment(dbms, config_path, use_cache)
-
+    start_db_container(dbms, cfg)
     start_sqlancer_container(
         dbms=dbms,
-        host_container_name=container_name,
+        host_container_name=cfg["container_name"],
         username=cfg["username"],
         password=cfg["password"],
         oracle=cfg["oracle"],
@@ -149,7 +100,7 @@ def test_single(dbms, config_path, use_cache=False):
         timeout=cfg["timeout_seconds"]
     )
 
-    remove_container(container_name)
+    remove_container(cfg["container_name"])
 
 def remove_container(container_name):
     try:
@@ -160,24 +111,14 @@ def remove_container(container_name):
 
 
 
-def test_all(use_cache=False):
-    global_cfg = load_json("config.json")
-    dbms_list = global_cfg.get("dbms_list", [])
-    for dbms in dbms_list:
-        config_path = os.path.join(dbms, "config.json")
-        if not os.path.exists(config_path):
-            print(f"[WARNING] Skipping {dbms}, missing config file.")
-            continue
-        test_single(dbms, config_path, use_cache)
 
-def test_custom_dockerfile(dockerfile_path, config_path, use_cache=False):
-    cfg = load_json(config_path)
-    dbms = cfg["dbms"]
-    cfg, container_name = build_environment(dbms, config_path, use_cache, True, dockerfile_path)
+def test_custom_dockerfile(dockerfile_path, cfg, use_cache=False):
+    dbms=cfg["dbms"]
 
+    start_db_container(dbms, cfg)
     start_sqlancer_container(
         dbms=dbms,
-        host_container_name=container_name,
+        host_container_name=cfg["container_name"],
         username=cfg["username"],
         password=cfg["password"],
         oracle=cfg["oracle"],
@@ -185,21 +126,6 @@ def test_custom_dockerfile(dockerfile_path, config_path, use_cache=False):
         timeout=cfg["timeout_seconds"]
     )
 
-    remove_container(container_name)
+    remove_container(cfg["container_name"])
 
-def ensure_network_exists(network_name="sqlancer-net"):
-    try:
-        output = subprocess.check_output([
-            "docker", "network", "ls",
-            "--filter", f"name={network_name}",
-            "--format", "{{.Name}}"
-        ])
-        networks = output.decode().splitlines()
-        if network_name not in networks:
-            print(f"[INFO] Creating docker network: {network_name}")
-            subprocess.run(["docker", "network", "create", network_name], check=True)
-        else:
-            print(f"[INFO] Docker network '{network_name}' already exists.")
-    except Exception as e:
-        print(f"[ERROR] Failed to check/create Docker network: {e}")
-        sys.exit(1)
+
